requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
et-xmlfile==1.0.1
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
jdcal==1.4.1
lxml==4.5.1
numpy==1.19.0
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.4
pandas==1.0.5
peewee==3.13.3
protobuf==3.12.2
pyactiveresource==2.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ShopifyAPI==8.0.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
